add code admin

+++ app/Http/Controllers/Admin/ProductController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 use App\Category;
 use App\Product;
+use App\Order;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 
@@ -15,8 +16,9 @@ class ProductController extends Controller
      */
     public function index()
     {
-        $listProducts = Product::all();
-        return view('admin.product.list', compact('listProducts'));
+        $products = Product::with('category')->orderBy('id','ASC')->paginate(2);;
+        $categoryID = Category::all();
+        return view('admin.product.list', compact('products','categoryID'));
     }
 
     /**
@@ -27,7 +29,7 @@ class ProductController extends Controller
     public function create()
     {
         $categoryID = Category::all();
-        return view('admin.product.create', compact('categoryID'));
+        return view('admin.product.create',compact('categoryID'));
     }
 
     /**
@@ -39,8 +41,15 @@ class ProductController extends Controller
     public function store(Request $request)
     {
         $data = $request->except('_token');
+        if($request->hasFile('image')){
+            //code upload file
+            $newName = md5(microtime(true)).$request->file('image')->getClientOriginalName();
+            $request->file('image')->move(public_path('admin/images/products/'),$newName);
+            
+        }
+        $data['image'] = '/admin/images/products/'.$newName;
         Product::create($data);
-        return redirect()->route('product.list');
+        return redirect()->route('product.index');
     }
 
     /**
@@ -51,7 +60,9 @@ class ProductController extends Controller
      */
     public function show($id)
     {
-        //
+        $product = Product::with('category')->find($id);
+        //dd($sp->category->name);
+        return view('admin.product.list-id', compact('product'));
     }
 
     /**
@@ -78,8 +89,15 @@ class ProductController extends Controller
     {
         $product = Product::find($id);
         $data = $request->except('_token', '_method');
+        if($request->hasFile('image')){
+            //code upload file
+            $newName =md5(microtime(true)).$request->file('image')->getClientOriginalName();
+            $request->file('image')->move(public_path('admin/images/products/'),$newName);
+            
+        }
+        $data['image'] = '/admin/images/products/'.$newName;
         $product->update($data);
-        return redirect()->route('product.list');
+        return redirect()->route('product.index');
     }
 
     /**
@@ -90,8 +108,40 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
-        $product = Product::find($id);
-        $product->delete();
-        return redirect()->route('product.list');
+        
+        $product = Product::with('order_details')->find($id);
+        $order_detail =$product->order_details;
+        foreach ($order_detail as $value) {
+            // lấy giá trị order_id//
+            $order_id = $value->order_id;
+            $order = Order::find($order_id)->toArray();
+            //lấy status_id//
+            $status_id = $order['status_id'];
+            //dd($status_id);
+            if($status_id == 4){
+                $product->delete();
+                return redirect()->route('product.index')->with(['message'=>'Đã xóa sản phẩm thành công. !!']);
+            }
+            else{
+                return redirect()->route('product.index')->with(['error'=>'Không thể xóa vì đơn hàng của sản phẩm này chưa thanh toán.!!']);
+            }    
+        }
+        
     }
+    public function search(Request $request){
+        $categoryID = Category::all();
+        if($request->name){
+            $products = Product::where('name','like','%'.$request->name.'%')
+                        ->orWhere('price',$request->name)
+                        ->get();
+        return view('admin.product.search', compact('products','categoryID'));
+        }
+        if($request->brand){
+            //dd($request->brand);
+            $products = Product::where('category_id',$request->brand)->get();
+                        
+        return view('admin.product.search', compact('products','categoryID'));
+        }
+    }
+    
 }
